requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
openai-whisper>=20231117
funasr>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
